feat(nsga3): implement NSGA-III backend runner with bug fixes
Implemented:
- NSGA3Runner class with clean API for multi-objective optimization
- NSGA3RunnerConfig for flexible configuration
- Best compromise solution selection (minimum sum of normalized objectives)
- Convenience function run_nsga3() for quick usage
- Statistics tracking (runtime, evaluations, pareto size)
- Support for custom and two-layer reference points

Bug Fixes:
1. Fixed import error: maximize_connectivity → maximize_network_connectivity
2. Fixed NSGA-III objectives array being overwritten by dict in evaluator
3. Fixed niche count dtype overflow (int → float to support np.inf)
4. Fixed Solution class to handle numpy array objectives in __init__ and copy()

Testing:
- Added 10 comprehensive unit tests for NSGA3Runner (all passing)
- Fixed 12 existing NSGA-III integration tests (all passing)
- Total: 155 unit + 28 integration tests

Files:
- New: backend/core/optimization/nsga3_runner.py (281 lines)
- New: backend/tests/unit/test_nsga3_runner.py (267 lines)
- New: DEVELOPMENT_ROADMAP.md (tracking project progress)
- Modified: 4 core files with critical bug fixes

Related to Phase 2.1 of the NSGA-III multi-objective optimization roadmap.

+++ src/algorithms/nsga3/niching.py
@@ -180,7 +180,8 @@ def niche_preserving_selection(
     n_ref_points = len(reference_points)
 
     # Count solutions associated with each reference point
-    niche_counts = np.zeros(n_ref_points, dtype=int)
+    # Use float dtype to allow np.inf
+    niche_counts = np.zeros(n_ref_points, dtype=float)
     for idx in front:
         niche_counts[associations[idx]] += 1
